New password hash, Default setup, Admin file access, No cgi
Changed the password hasher to the blowfish model now suggested by
many. Got file access again for admin. Created a sort of default
config scheme where the system has not config file to start with,
and forces the admin to make one, as well as set the admin password.
Maybe all cgi functions are now unused. We'll have to see as we
move some files.