Use symlinks to read the files.
The NDAS might mount to a different location than the web root.
A way to get the files is to use a symlink in a web accessible dir.
It need testing, but for now, admin can get the files.
If it works, it will have a folder for each user to download files
from.
Index.php in the files/ folder is hoping to keep surfers out of the
links that remain.