Try to mount RO NTFS with errors. Fixed RW NTFS false error.
It is OK to mount some NTFS partitions RO even if there is some
error, so on the RO NDAS especially, we can try to mount them.

The writeable check function was sending back the device name
with the RW indicator if the partition is ntfs. Now it only sends
back RW.